Вынесен TasksLinkedList в отдельный файл, т.к. Github Actions требовал это при проверке CheckStyle
