+++ src/managers/InMemoryHistoryManager.java
@@ -2,7 +2,6 @@ package managers;
 
 import tasks.Task;
 
-import java.util.ArrayList;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -39,55 +38,3 @@ public class InMemoryHistoryManager implements HistoryManager {
     }
 }
 
-class TasksLinkedList<T> {
-    public Node<T> head;
-    public Node<T> tail;
-    private int size = 0;
-
-    public int getSize() {
-        return size;
-    }
-
-    public void linkLast(T task) {
-        final Node<T> oldTail = tail;
-        final Node<T> newNode = new Node<T>(oldTail, task, null);
-        tail = newNode;
-        if (oldTail == null)
-            head = newNode;
-        else
-            oldTail.next = newNode;
-        size++;
-    }
-
-    public void removeNode(Node<T> node) {
-        Node<T> prev = node.prev;
-        Node<T> next = node.next;
-
-        if (prev == null) {
-            head = next;
-        } else {
-            prev.next = next;
-            node.prev = null;
-        }
-
-        if (next == null) {
-            tail = prev;
-        } else {
-            next.prev = prev;
-            node.next = null;
-        }
-
-        node.data = null;
-        size--;
-    }
-
-    public List<T> getTasks() {
-        List<T> history = new ArrayList<>();
-        Node<T> node = head;
-        while (node != null) {
-            history.add(node.data);
-            node = node.next;
-        }
-        return history;
-    }
-}
